rename soft_exceptions to soft_instructions
