+++ sbi/ogsbi/other_exceptions.c
@@ -16,7 +16,7 @@
 
 #include "asm_c_interface.h"
 #include "common.h"
-#include "soft_exceptions.h"
+#include "soft_instructions.h"
 
 #include <assert.h>
 #include <stdio.h>
@@ -55,7 +55,7 @@ void handle_other_exceptions(__attribute__((unused)) uint32_t registers[31], __a
             break;
         case 2://ILLEGAL_INSTRUCTION_EXCEPTION
             dputs("Illegal Instruction Exception");
-            soft_exception_handler(mepc, registers);
+            soft_instruction_handler(mepc, registers);
             //assert(false && "TODO implement (emulate some instructions, and delegate others to S-Mode)");
             break;
         case 3://BREAKPOINT_EXCEPTION
